refactor: 아이콘 currentColor 기반으로 통일 (#154)
